refactor: GA 위치 변경 (#499)
