+++ frontend/src/router.tsx
@@ -5,8 +5,6 @@ import RootPage from './pages/RootPage';
 import { ReactNode } from 'react';
 import AuthLayout from './components/Layout/AuthLayout';
 import NotFound from './pages/NotFound';
-import ReactGA from 'react-ga';
-import { createBrowserHistory as createHistory } from 'history';
 
 const SelectedTopic = lazy(() => import('./pages/SelectedTopic'));
 const NewPin = lazy(() => import('./pages/NewPin'));
@@ -19,13 +17,6 @@ const SeeTogetherTopics = lazy(() => import('./pages/SeeTogetherTopics'));
 const Profile = lazy(() => import('./pages/Profile'));
 const AskLogin = lazy(() => import('./pages/AskLogin'));
 const Bookmark = lazy(() => import('./pages/Bookmark'));
-
-ReactGA.initialize('G-RRSTX6Y61Y', { debug: true });
-const history = createHistory();
-history.listen((location: any) => {
-  ReactGA.set({ page: location.pathname });
-  ReactGA.pageview(location.pathname);
-});
 
 interface routeElement {
   path: string;
